Updated layout and style, optimized code structure
In this commit, we made a series of code changes to optimize the layout and style of the project. First, we added some files to the `.gitignore` file so that Git would ignore them. Then, we reordered the HTML tag attributes in the `App.razor` file and added new link and script tags. We also added and rearranged some CSS rules in the `MainLayout.razor.css` file. In addition, we added the `TopNavBar.razor` and `TopNavBar.razor.css` files, which define a new top navigation bar component and its style. In the `Routes.razor` file, we reformatted the route definitions to improve readability. In the `libman.json` file, we changed the name of the library and removed the reference to the `bootstrap` library. Finally, we added the `tailwind.config.js` and `tailwind.extension.json` files, configured the Tailwind CSS, and added references to the Tailwind CSS and some new CSS rules in the `app.css` file.