Added new links and interface elements to TopNavBar and Home components!
Some important changes have been made in the `TopNavBar.razor` and `Home.razor` files. In `TopNavBar.razor` we added new constant `docsLink` and updated the links to the top navigation bar and mobile device menu. In `Home.razor` we have added a new tab navigation below the `OTC Current Status` tab and a new area at the bottom of the page with three different coloured objects.