Updated Home.razor styles and components, added StatusCard component
In the `Home.razor` file, we have made some changes to the style of the label for OTC's current status, including font size, thickness and colour. At the same time, we removed the three existing objects and replaced them with four new `StatusCard` components. To improve responsiveness, we also adjusted the spacing and number of columns in the grid layout.

In addition, we added a new component called `StatusCard`, which is defined in the `StatusCard.razor` file. This component contains a title and two list items, each preceded by an icon indicating success.