Updated UI components and styles
In the `StatusCard.razor` file, the code listing the status of the service has been removed in favour of using the new `ServiceItem` component to display the status of the service. The `ServiceItem` component accepts `Name` and `Type` parameters to display the name and status type of the service.

In the `PageFooter.razor` and `Home.razor` files, new classes have been added to several elements to change the elements' margins and internal layout.

Added the `ServiceItem.razor` file, which defines the new `ServiceItem` component for displaying the status of a service.