Updated code for Indicator, StatusCard and Home components.
In Indicator.razor.cs, removed the fillColor property and integrated its logic into the icon property, and updated the name and use of the pref constant. In StatusCard.razor, added classes to div and scale-divider tags and changed the generation of ServiceItem tags to a loop. In Home.razor, removed useless text, updated the class for div tags, and added a loop to generate div tags.