Added field constants, field options, page and component property settings
Added `FieldTypes`, `SortBy` and `TextAlign` records defining string constants for field types, sorting methods and text alignment. In `FieldOption.cs`, a new `FieldOption` record has been added that defines the properties of the field options and specifies the JSON property names for each property. In `StatusCard.razor`, optimised the way the `Type` property of the `ServiceItem` component is used. In `Dev.razor`, added a new `Dev` page and set the properties of the `scale-data-grid` component.