Updated DataGrid components and field options.
In the `DataGrid.razor` file, removed the injection of `IJSRuntime JS` and added a reference to `StatusDashboard.Components.Home`. Also, added a `foreach` loop that iterates over all the values of the `StatusType` enumeration and creates a template for each value that contains an `Indicator` component of type current enumeration value.

In the `DataGrid.razor.cs` file, a reference to `Home` is first added to the namespace. Then, the field types in the `fields` list were modified and in the `OnAfterRenderAsync` method, a new `module` was created if it was rendered for the first time and the `setFields` method was called to set the fields. Also, a new `rows` list is added and the `setRows` method is called to set the rows.

In the `DataGrid.razor.ts` file, the `dataGrid.rows` and `dataGrid.localisation` settings have been removed and a new `setRows` function has been added to set the rows.

In the `FieldConst.cs` file, changed `FieldTypes`, `SortBy` and `TextAlign` from `record` type to `static class` type.

In the `FieldOption.cs` file, added a new `FieldOption` record with a required `Type` attribute.