Optimised DataGrid component code and enhanced functionality
In the `DataGrid.razor` file, optimised the formatting of the `<scale-data-grid>` tag to improve code readability. In the `DataGrid.razor.cs` file, removed the `size` property of an object. In the `DataGrid.razor.ts` file, the `CurrentEvent` property of the global interface `Window` has been removed, a new constant `dataGrid` has been added, and a number of operations have been performed on it, including hiding the borders, observing changes in the child nodes and processing them, and optimising the formatting of the arguments to the `setRows` function. These changes are mainly code formatting optimisations and enhancements to the `DataGrid` component.