Updating DataGrid processing and adding new pages
In the `DataGrid.razor.cs` file, changes were made to the `DataGrid` class, including changing field labels, adding a non-null assertion operator, and adjusting the initialisation of data objects. In the `DataGrid.razor.ts` file, optimisations were made to the handling of the data grid, including changing `dataGrid` from a constant to a variable, creating a new `refresh` function, and tweaks in the `setFields` and `setRows` functions. In addition, new pages have been added to the `History.razor` file.