Updating components and pages to optimise dependencies
In `EventItem.razor`, we changed the formatting of the datetime and changed the background colour of the `Indicator` component from `bg-slate-100` to `bg-zinc-100`. In `EventSection.razor`, we have replaced the `Event` component with the `EventItem` component. In `MainLayout.razor`, we changed the background colour of `Container` and removed the corresponding background colour setting in `MainLayout.razor.css`. In `Event.razor`, we added a new page that receives an integer parameter called `id`. In `History.razor`, we changed the page's routing from `/history` to `/History`. In `package.json`, we moved `@types/blazor__javascript-interop` from `devDependencies` to `dependencies` and removed the dependency on `@telekom/scale-components`. Also, we added new `devDependencies`: `@types/blazor__javascript-interop`.