Updated Indicator and Event components, added EventEditor component
In the `Indicator.razor.cs` file, the `Type` parameter has been changed and the `EditorRequired` property has been added to make it a required parameter in the editor. In the `Event.razor` file, a new namespace `StatusDashboard.Components.Event` has been introduced and the layout of the `Indicator` component has been adjusted and the `EventEditor` component has been added. In addition, the layout of the header was modified and some `div` tags were removed, while some tags containing event details were added to the new `div` tags. Finally, in the new `EventEditor.razor` file, a modal dialog containing a welcome message and a button to open this dialog were added.