Updated form styling and validation, added new development dependencies
In the `EventEditor.razor` file, we added new CSS classes to change the layout and spacing of the form, tweaked the style classes for the input elements, changed the colour of the validation message, and improved the user experience. In the `EventForm.cs` file, we removed the data annotation for the `Type` property. In the `package.json` file, we added a new development dependency `"@tailwindcss/forms"` and adjusted the order of the development dependencies. Finally, in the `tailwind.config.js` file, we enabled the newly added Tailwind CSS forms plugin.