Updating EventEditor component and related code
In `EventEditor.razor` we have updated the `EditForm` component by adding the `OnValidSubmit` property and pointing to the new `submit` method. We removed the `InputNumber` component and changed the way the `EventType` enumeration value is traversed in the `InputSelect` component. We also added a new `InputSelect` component for binding to `model.Status` and added a `scale-button` component after the `ValidationMessage` component.

In `EventEditor.razor.cs`, we removed the `SupplyParameterFromForm` feature from the `model` property and added a non-null check for `module` in the `DisposeAsync` method. We have simplified the `openModal` method and added initialisation of `module` in the `OnAfterRenderAsync` method. We also added a new `submit` method that prints `model.Title` to the console.

In `EventForm.cs`, we removed the required `EventId` property.