Update label properties in EventEditor.razor
In the EventEditor.razor file, we updated the `for` and `id` attributes of `<label>` and `<InputText>`, `<InputSelect>`, `<InputTextArea>` and other labels. Now, these properties are no longer hard-coded strings, but use `@nameof(model.PropertyName)`. In this way, if the attribute name of the model changes, the related labels will be automatically updated, thus improving the maintainability of the code.