Updating the events page and adding new components
In the `Event.razor` file, we removed the existing code for displaying event details and introduced two new components, `<EventCard />` and `<EventLog />`. The `EventCard` component is responsible for displaying event details, while the `EventLog` component is responsible for displaying event log information. Also, we have kept a parameter named `Id` for receiving and processing the ID of the event.In the two new files `EventCard.razor` and `EventLog.razor`, we have added the code for displaying the event details and the event log respectively.