Updated region selector and event status, added availability matrix
In the `AvailaMatrix.razor` file, added a new table that shows the availability percentages for different services for different months. In the `EventCard.razor` file, changed the fixed text "Fixed" to the dynamic `@EventStatus.Resolved`. In the `EventLog.razor` file, removed the `show-sort` attribute and removed the `aria-disabled="true"` attribute from the table header. In the `RegionSelector.razor` file, added a new region selector component that contains a header and two tabs "EU-DE" and "EU-NL". In the `RegionSelector.razor.cs` file, added a new parameter `Title` to the `RegionSelector` component. In the `RegionSelector.razor.css` file, new styles have been set for `#RegionSelector` and `scale-divider`. In the `Availability.razor` file, added a new page that uses the `RegionSelector` and `AvailaMatrix` components. In the `Home.razor` file, replaced the old region selector section with a new `RegionSelector` component and updated the title. In the `Home.razor.css` file, removed the style settings for `#RegionSelector` and `scale-divider`.