Updated variable names and components, added NavItem component and its backend logic
In the `App.razor` file, we changed the variable name from `unpkg` to `cdn` and used it to link CSS styles and JavaScript modules. In the `TopNavBar.razor` file, we removed the `scale-telekom-nav-item` component and replaced it with a new `NavItem` component that accepts `href` and `Label` as parameters. Also, we have replaced the `NavLink` component with an `a` label for creating navigation links in mobile menus. In addition, we have added new `NavItem.razor` and `NavItem.razor.cs` files that define the front-end presentation and back-end logic of the `NavItem` component, respectively.