Updated form fields and validation, optimised datetime formatting
In the `EventItem.razor` file, the datetime format was changed and the background colour of the indicators was modified. In the `NewForm.razor` file, a number of changes have been made, including the use of new input fields and selectors, the addition of validation messages for `model.Description` and `model.Start`, and the display of a date and time selector for `model.End` under certain conditions. In the `NewModel.cs` file, a number of changes were made to the `NewModel` class, including setting default values and adding properties. Finally, a new `DateTimeBindAttribute` class has been added to validate the date-time order.