Updated form styles, optimised model and service selectors
In the `NewForm.razor` file, adjusted the `class` property of the `EditForm` class and optimised the vertical inner margins of the form. In the `NewModel.cs` file, updated the type definition of the `Services` property of the `NewModel` class and added a minimum length annotation. In the `ServiceSelector.razor` file, a number of changes were made, including updating the type and initialisation of the `items` list, optimising the column definitions for the `FluentDataGrid`, and removing class definitions that were no longer needed. In the `ServiceSelector.razor.cs` file, new class definitions have been added.