Adding new StatusService and StatusOption classes and updating the configuration
In `Program.cs`, we added a reference to `StatusDashboard.Services` and added some services to the container. These services include configuration options for `StatusOption`, an HTTP client, and a single instance of `StatusService`.

In `StatusService.cs`, we removed the `StatusService` class.

In `Status.Entities.cs`, we changed the namespace from `StatusDashboard.Services.ComponentStatus` to `StatusDashboard.Services`.

In `Status.Option.cs`, we added a new `StatusOption` class under the `StatusDashboard.Services` namespace that contains a required `Source` property of type `URL` with a minimum length of seven.

In `Status.cs`, we added a new `StatusService` class under the `StatusDashboard.Services` namespace that contains a `logger` property of type `ILogger<StatusService>`, a `StatusOption` property of type `StatusOption`, and a `Source` property of type `StatusOption`. StatusOption` property of type `option`, and a `httpClient` property of type `HttpClient`. In addition, a constructor has been added that accepts an `ILogger<StatusService>` parameter, a `IOptions<StatusOption>` parameter, and an `IHttpClientFactory` parameter, and initialises the properties using these parameters.

In `appsettings.json`, we add a new `SDB` configuration that contains a `Source` property with a value of "https://status.otc-service.com/".