Adding StatusService as a Managed Service
In the Program.cs file, we added a line of code to register StatusService as a hosted service. We also changed the definition of the StatusService class so that it implements the IHostedService interface and added implementations of the StartAsync and StopAsync methods. These changes mean that StatusService will start when the application is started and stop when the application is closed.