Updated status entities and service classes, added new enum values and handlers, optimised date types and JSON converters
In the `Status.Entities.cs` file, we modified the date type properties `EndDate`, `StartDate` and `Timestamp` of the `IncidentEntity` and `UpdateEntity` classes so that they can now accept null values. Also, we added two new enumeration values `Scheduled` and `Fixing` to the `StatusEnum` enumeration class and added handling of these two new enumeration values in the `Read` method of the `StatusEnumConverter` class. In addition, we have changed the type of the `UtcConverter` class from `JsonConverter<DateTime>` to `JsonConverter<DateTime?>` and the return type of the `Read` method from `DateTime` to `DateTime?`, meaning that ` UtcConverter` can now handle null date values.

In the `Status.cs` file, we declared the `StatusService` class as a partial class and added a new private method, `getStatus`, which gets status information from the specified endpoints and prints out the name of each status. The `StartAsync` method now calls the new `getStatus` method instead of just printing "Start".

Finally, in the `Status.DB.cs` file, we added a partial definition of the `StatusService` class.