Updated definitions and properties of the model class and `StatusContext` class
In the `StatusContext` class, changed it from `public` to `internal` and added five new `DbSet` properties. In `Status.Models.cs`, all model classes have been changed from the `Model` suffix to no suffix, and all `required` keywords have been removed. The `StringLength` attribute has been added to the `Name` property in all model classes, limiting it to a maximum length of 255 and a minimum length of 1. The type of the collection attribute in the `Service`, `Category`, and `Region` classes has been modified. New properties have been added to the `Event` and `History` classes, and the `StringLength` property has been added to some properties to limit the length. A `#pragma warning disable CS8618` has been added to all model classes to disable the compiler from generating a warning that non-null fields must contain non-null values when exiting the constructor.