Update the usage of DbContext and HttpClient, optimise logging settings, add StatusHttp class.
In `Program.cs`, changed `AddDbContext` to `AddDbContextFactory` and `AddHttpClient` to `AddHttpClient<StatusHttp>` to use the DbContext factory and the specific HttpClient type `StatusHttp`. In `Status.cs`, updated the constructor of the `StatusService` class by replacing `IHttpClientFactory` and `StatusContext` with `IDbContextFactory<StatusContext>` and `StatusHttp` and moving the functionality of the ` getStatus` method's functionality to the `StartAsync` method. In `appsettings.Development.json` and `appsettings.json`, adjusted logging level settings and added logging for `Microsoft.EntityFrameworkCore`. Added `Status.Http.cs` file that defines the `StatusHttp` class, which contains an instance of HttpClient and a method `GetStatus` to get status.