Updated database model and service operations
In the `Status.Context.cs` file, a new `DbSet` called `EventService` was added to the `StatusContext` class. In the `Status.Models.cs` file, changes were made to the `Service`, `Region`, and `Event` classes, and new `Category` and `EventService` classes were added. These changes include adding a reference to `Microsoft.EntityFrameworkCore`, adding a unique index to the `Name` property, adding collection types to the `Services` and `Histories` properties, and adding a new property `Abbr`.

In the `Status.cs` file, changes have been made to the constructor and `StartAsync`, `StopAsync` methods of the `StatusService` class, including the removal of assignments to `logger`, `option`, `context`, and `http`, and the addition of operations on databases, including operations on ` Categories`, `Regions`, `Services`, removed printing of `item?.Name`, added save operations to the database, and changed the original methods to expression body members.