Enhanced event status and history handling
In `EventEditor.razor`, changed the way `status` is assigned so that it now skips the first four enumeration values and then takes the next four. In `EventStatus.cs` and `Status.Entities.cs`, added new enumeration values `SysInfo` and `Observing` to the `EventStatus` and `StatusEnum` enumerations, respectively. In `Status.Models.cs`, added new properties `Status` and `Event` to the `History` class. In `Status.cs`, a reference to `Components.Event` has been added and a number of modifications have been made to the `StatusService` class, mainly adding the creation and assignment of `Event` and `History` to each `incident` as it is processed. These modifications mainly enhance the handling of event status and history by adding new status types and creating the corresponding events and histories when an event is processed.