Update Home.razor and Home.razor.cs to dynamically display exceptions
In Home.razor, changed the values of the `heading` and `variant` properties from hard-coded to dynamic to change dynamically based on the number of exceptions. In Home.razor.cs, imported the `Components.Event` namespace and added the `abnormalCount` and `heading` properties to count and display the number of events in the database that are open and of a type other than maintained. An asynchronous release to the database has also been added to the `DisposeAsync` method. These changes were made primarily to allow the page headings and warning levels to dynamically change based on the number of abnormal events in the database.