Enhance the interoperability of EventItem components
Introduced JavaScript interop functionality to `EventItem.razor`, allowing Razor components to interact with JavaScript code through the `IJSRuntime` service. Modified the HTML representation of the event start date from a `<li>` element to a quoted `<label>` element so that JavaScript can directly manipulate this element. In `EventItem.razor.cs`, added a reference to `Microsoft.JSInterop` and a private `labelElement` property to store a reference to the front-end `<label>` element, and overridden the `OnAfterRenderAsync` method , to call the `onLabel` function in the JavaScript module after the component is rendered for the first time. In addition, the type of `Item` property is simplified, and the `onLabel` function is defined in the newly added `EventItem.razor.ts` file, which is used to adjust the style of related elements to improve layout and visual effects.