Refactoring EventItem component logic
In `EventItem.razor`, references to components that are no longer used have been removed, code to retrieve information directly from the database has been removed, and logic to set colors based on the state of an event has been removed. These changes simplify the front-end code and reduce direct interaction with the database.

In `EventItem.razor.cs`, new private properties have been added and logic from the previous `.razor` file has been moved to the back-end, including logic that handles service and region information associated with events and sets colors. This step marked a migration of logic from the front-end to the back-end and was intended to improve code maintainability and efficiency of data handling.

In addition, minor adjustments were made to the use of the `<scale-tag>` component to fix a formatting issue when displaying the number of additional services.