Optimizing the dynamic display of the Event Log component
Important changes have been made to the `EventLog.razor` and `EventLog.razor.cs` files to enhance the dynamic display of the event log component. In `EventLog.razor`, the static display of event statuses and messages has been removed in favor of dynamically reading and displaying the status, UTC creation time, and messages of the event history from the `list` collection via a `foreach` loop. In addition, a reference to `Microsoft.EntityFrameworkCore` has been added to the `EventLog.razor.cs` file and a new `list` property has been introduced to store all history records retrieved from the database that are associated with an event ID. Retrieval of event history and sorting in descending order by creation time is achieved through Entity Framework Core's LINQ queries, which improves the flexibility and usability of the component.