Updated logic for displaying categories and services
In `AvailaMatrix.razor`, updated the way the table is populated to create `<CategoryGroup />` components for each category by looping through them and passing the `Category` parameter. In `AvailaMatrix.razor.cs`, improved the way the `categories` variable is fetched by adding sorting by name. In the `CategoryGroup.razor` file, several changes have been made to the HTML structure of the component, including dynamic display of category names, changes to the logic for displaying service names, updates to the table cell color class, and modifications to the way the list of services is displayed. In `CategoryGroup.razor.cs`, added `Category` and `Region` attributes to the `CategoryGroup` component and marked them as required, added a new `services` array to store the services, and updated the logic for getting the services in `OnParametersSetAsync` , including sorting by service name.