Optimized code and introduced SLA service calculation functionality
The `AvailaMatrix.razor.cs` file has been simplified to simplify the code structure by replacing the curly brackets in the `OnInitializedAsync` and `OnParametersSetAsync` methods with arrow functions. At the same time, the logic of setting the `categories` property by database query in the `OnParametersSetAsync` method remains unchanged.

In the `CategoryGroup.razor` file, the `SlaService` service has been introduced and utilized to display the calculated SLA values in the Razor component, replacing the cell contents that were fixed at `100.00`. This change is reflected in the display of the first and subsequent entries of the service.

In the corresponding `CategoryGroup.razor.cs` file, a private property named `slas` has been added to store the list of calculated SLA values. In the `OnInitializedAsync` method, the `Calc6Months` method of `SlaService` is called for each service to calculate the SLA values for the last six months and the results are added to the `slas` property.

In the `Program.cs` file, removed some unused `using` directives and added `SlaService` as a singleton service to the service container to support the above functionality.

Added a new `SLA.cs` file containing the `SlaService` class. This class is responsible for calculating the SLA value for the service, calculating the SLA value for each month of the last six months for the specified service via the `Calc6Months` method, excluding maintenance-type events, and taking into account the total number of minutes and downtime for each month to calculate the SLA. in addition, the `SlaService` class implements the `IAsyncDisposable` interface to ensure that the database context is released correctly. database context is released correctly.