Introducing memory caching to improve performance
A memory cache service has been added to `Program.cs` to allow applications to use the memory cache to store data, thereby increasing the speed and efficiency of data access. The `SlaService` class in the `SLA.cs` file has been modified to add a parameter of type `IMemoryCache` to allow it to utilize the memory cache. In addition, the `Calc6Months` method has been added to the `SlaService` class to calculate the last six months of data and attempt to fetch the results from the cache, or if there is no cache, calculate and store the results with a cache time of 1 hour. Also, the logic for calculating the total downtime per month has been optimized to use a LINQ query instead of the original round-robin calculation method, and the results are stored in the memory cache to further improve performance. These changes are intended to improve the efficiency of the application in processing data and performing calculations by introducing memory caching.