Optimized EventGrid display logic, updated configuration
Several important updates have been made to the `EventGrid.razor.cs` file, the first being the addition of a TODO note for future database migrations, suggesting that when using PGSQL databases, consideration should be given to migrating the relevant code to LINQ.Additionally, two new variables `regions` and `services` have been introduced, optimizing the de-duplication and array conversion logic for regions and services by applying the `Distinct(.ToArray()` method to them. By applying the `Distinct().ToArray()` method to them, the de-duplication and array conversion logic for regions and services has been optimized. This change also affects the presentation of Regions and Services in the returned object array, improving the presentation when the number of Regions or Services is greater than 1 by replacing the original presentation logic with de-weighted data.

At the same time, the `appsettings.Production.json` file has been modified to add a new `"urls": "http://*:3389"` configuration entry, which enables the app to listen on port 3389 on all network interfaces in the production environment, an important configuration update to support wider network access.