Enhancements to Indicator components and layout adjustments
The `System.ComponentModel.DataAnnotations` namespace has been introduced in `Indicator.razor.cs` to support data annotations. A new optional parameter named `Size` has been added to the `Indicator` component to allow setting the icon size, limited to a range of 12 to 128. In addition, when the `Size` parameter is not null, the `icon` rendering fragment adds a corresponding `size` attribute based on the `Size` value, which directly affects the rendered output.

In `ServiceItem.razor`, a reference to the `StatusDashboard.Components.Event` namespace has been added and the component display logic has been modified. The new `future` private field and its logic update in the `OnParametersSetAsync` method enables the component to decide whether to display special layouts (containing `FluentCounterBadge` and `Indicator`) based on the start time of the event compared to the current UTC time to distinguish between the upcoming events and the current state.

Changes in the `Home.razor` file include the addition of `flex-wrap` and `gap-y-2` classes to the two `section` elements, designed to improve responsive layouts and ensure that content aligns better across different screen sizes.