Enhancements to the application's authentication mechanism
Changes have been made to the `ProfileMenu.razor`, `NewEvent.razor`, and `Routes.razor` files to enhance the application's user authentication features. In `ProfileMenu.razor`, the login link has been removed, the `Microsoft.AspNetCore.Components.Authorization` namespace has been introduced, and the `AuthorizeView` component has been added to distinguish between authorized and unauthorized users. The `[Authorize]` property has been added to the `NewEvent.razor` file to require the user to log in to access the page and introduce the necessary namespaces. The `Routes.razor` file enables authentication state to be passed in the application by introducing the `Microsoft.AspNetCore.Components.Authorization` namespace and wrapping the `RouteView` component with `CascadingAuthenticationState`. Authentication state is passed in the application.

In the `Program.cs` file, support for the Controller and Fluent UI components is added and the application startup is configured to use the authentication mechanism. Added `"EnableRolesMapping": "Realm"` configuration item to the `appsettings.json` file for role mapping policies.

A new `AuthController.cs` file was added to create an `AuthController` controller that contains login and logout actions and supports OpenID Connect and cookie-based authentication schemes. Together, these changes enhance the application's management and control of user login state, ensuring that authentication state is passed and used correctly.