Enhanced web application security and compatibility
Added new service configurations to `Program.cs`, primarily including configuration of `CookiePolicyOptions` to ensure that all users need to agree to cookie use, and setting `MinimumSameSitePolicy` to `SameSiteMode.Unspecified` to support cross-site to support cross-site requests carrying cookies, with possible security implications. At the same time, OpenID Connect support has been introduced, which uses Keycloak as an authentication service to enhance the security of the application and the user authentication process.

In addition, for non-development environments, HTTP requests are forced to redirect to HTTPS via `app.UseHttpsRedirection();` and `app.UseCookiePolicy();` has been added to ensure that the application follows the configured cookie policy, especially with regard to handling user consent and cross-site requests.