Updated App.razor and Program.cs configuration 🤣👉
In the App.razor file, we added two new HTML elements to improve the site's SEO and user experience. First, a <meta> tag was added to provide a description of the site's content, which reads "Your go-to resource for monitoring the availability of various components in different areas." Second, a <title> tag was introduced to set the site title to "Open Telecom Cloud Status Dashboard" to more accurately reflect the purpose and content of the site.

In the Program.cs file, we modified the configuration of the static file service for non-development environments to improve the cache management of the website. The original app.UseStaticFiles() call was removed and a new app.UseStaticFiles(new StaticFileOptions {...}) call. With this change, a "Cache-Control" header has been added to all static file responses with the value set to "no-cache, max-age=2592000". This configuration is designed to balance cache control and performance gains by instructing clients not to cache these files, while allowing the files to be on the server for a maximum age of 30 days.