Updates to `NewForm.razor.cs` and `NewModel.cs` files
In `NewForm.razor.cs`, the `submit` method has been significantly changed from synchronous to asynchronous, and logic has been added to handle `regions` and `targets`. In addition, the `NewModel` class in `NewModel.cs` has been modified to remove the minimum length requirement and property declarations for the `Services` property in order to simplify the model or to cope with situations where `Services` are no longer needed.