Updated development environment configuration to remove default configuration
Added `StatusDashboard` and `Keycloak` configuration items to `appsettings.Development.json` for the status dashboard and Keycloak authentication service configuration respectively. The `Source` for `StatusDashboard` is set to `http://80.158.108.251:5000/`, while the `Keycloak` configuration consists of `realm`, `auth-server-url`, `ssl-required`, `resource`, ` secret` under `credentials` and many other sub-configuration items, aiming to provide more flexible authentication service settings.

At the same time, the corresponding `StatusDashboard` and `Keycloak` configuration items, which previously contained the same settings as the development environment, have been removed from `appsettings.json`. This change is intended to clean up the default configuration files and ensure that these specific services and features are not enabled in production or other environments, in order to enhance the security and suitability of the application.